refactor: adjust imports, clean up hazards layer config
